Extract the top open-api specs code from Monitoring
